COLL BASE: add support for ompi_comm_coll_ignore and ompi_comm_coll_request info keys
ompi_comm_coll_request can be used to request a specific coll component to be chosen from the list of available components. If the component is not available the operation will fail.

ompi_comm_coll_ignore can be used specify a list of coll components to be ignored during the selection process

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>