COLL: Add ompi_comm_coll_preference info key to control collective component selection
The info value is a comma-separated list of entries, which are chosen with
decreasing priorities. This overrides the priority of the component,
unless the component has disqualified itself.
An entry prefixed with ^ starts the ignore-list. Any entry following this
character is removed from the list of available components.

Example: "sm,libnbc,^han,adapt" gives sm the highest preference, followed
by libnbc. The components han and adapt are ignored in the selection process.

Signed-off-by: Joseph Schuchart <schuchart@icl.utk.edu>